implement job aquisition by id and status

+++ server/server.go
@@ -32,7 +32,7 @@ type Client interface {
 type JobService interface {
 	GetJobs() ([]Job, error)
 	GetJobsByStatus(status string) ([]Job, error)
-	GetJobByID(jobID int) (Job, error)
+	GetJobByID(jobID int64) (Job, error)
 	CreateJob(m *Job) error
 	CreateJobs(job []Job) ([]Job, error)
 	AcquireNextQueuedJob() (Job, error)
